pyproject.toml
[project]
name = "alphavantage"
version = "0.3.8"
description = "Alphavantage MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bump2version>=1.0.1",
    "load-dotenv>=0.1.0",
    "mcp>=1.0.0",
    "toml>=0.10.2",
]
[[project.authors]]
name = "Cesar Alvernaz"
email = "cesar.alvernaz@gmail.com"


[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alphavantage_mcp_server"]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.26.0",
    "ruff>=0.9.9",
]

[tool.pytest.ini_options]
markers = [
    "asyncio",
]

[project.scripts]
alphavantage = "alphavantage_mcp_server:main"
